requirements.txt
Babel==2.6.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
coverage==4.5.2
CryptoCoinChartsApi==0.0.1
Django==1.7
django-admin-bootstrapped==2.5.7
django-appconf==1.0.2
django-bitcoin==0.2
django-common-helpers==0.9.2
django-crispy-forms==1.7.2
django-cron==0.5.1
django-dbsettings==0.10.0
django-extensions==2.1.4
django-gravatar==0.1.0
django-guardian==1.4.9
django-ipware==2.1.0
django-mailgun==0.9.1
django-nexmo==1.0.1
django-notification==1.3.3
django-phonenumber-field==2.1.0
django-session-security==2.6.2
django-social-widgets==0.4.0
django-sslify==0.2.7
django-tinymce==2.7.0
django-twilio==0.9.3
django-user-accounts==2.1.0
django-webtest==1.9.4
djangorestframework==3.9.0
djutils==0.3.2
ez-setup==0.9
fake-useragent==0.1.11
filelock==3.0.10
html5lib==1.0.1
huey==1.10.5
idna==2.7
MySQL-python==1.2.5
oauthlib==2.1.0
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
PyJWT==1.6.4
pyPdf==1.13
PyPDF2==1.26.0
python-magic==0.4.15
python-memcached==1.59
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2018.7
PyYAML==3.13
redis==3.0.1
reportlab==3.5.12
requests==2.20.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-core==2.0.0
suds==0.4
toml==0.10.0
tox==3.6.1
twilio==6.22.0
typing==3.6.6
urllib3==1.24.1
virtualenv==16.1.0
waitress==1.1.0
webencodings==0.5.1
WebOb==1.8.4
WebTest==2.0.32
xhtml2pdf==0.2.3
xmltodict==0.11.0
